requirements.txt
Django==5.0.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pillow==10.2.0
django-cors-headers==4.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
django-filter==25.1
pytest==8.3.5
pytest-django==4.10.0
pytest-cov==6.1.0
